+++ testkit/index.js
@@ -12,9 +12,6 @@ export const tooltipTestkitFactory = testkitFactoryCreator(tooltipDriverFactory)
 import backofficeTooltipDriverFactory from '../src/Backoffice/Tooltip/Tooltip.driver';
 export const backofficeTooltipTestkitFactory = testkitFactoryCreator(backofficeTooltipDriverFactory);
 
-import labelDriverFactory from '../src/Label/Label.driver';
-export const labelTestkitFactory = testkitFactoryCreator(labelDriverFactory);
-
 import breadcrumbsDriverFactory from '../src/Breadcrumbs/Breadcrumbs.driver';
 export const breadcrumbsTestkitFactory = testkitFactoryCreator(breadcrumbsDriverFactory);
 
@@ -205,5 +202,6 @@ export {
   stylableCounterBadgeTestkitFactory as counterBadgeTestkitFactory,
   headingTestkitFactory,
   toggleSwitchTestkitFactory,
-  textTestkitFactory
+  textTestkitFactory,
+  labelTestkitFactory
 } from 'wix-ui-backoffice/dist/src/testkit';