+++ testkit/protractor.js
@@ -12,9 +12,6 @@ export const richTextAreaTestkitFactory = protractorTestkitFactoryCreator(richTe
 import inputAreaDriverFactory from '../src/InputArea/InputArea.protractor.driver';
 export const inputAreaTestkitFactory = protractorTestkitFactoryCreator(inputAreaDriverFactory);
 
-import labelDriverFactory from '../src/Label/Label.protractor.driver';
-export const labelTestkitFactory = protractorTestkitFactoryCreator(labelDriverFactory);
-
 import buttonDriverFactory from '../src/Backoffice/Button/Button.protractor.driver';
 export const buttonTestkitFactory = protractorTestkitFactoryCreator(buttonDriverFactory);
 
@@ -133,5 +130,6 @@ export {
   stylableCounterBadgeTestkitFactory as counterBadgeTestkitFactory,
   headingTestkitFactory,
   toggleSwitchTestkitFactory,
-  textTestkitFactory
+  textTestkitFactory,
+  labelTestkitFactory
 } from 'wix-ui-backoffice/dist/src/testkit/protractor';